[To dev/1.3] Pipe: Reduced the progress index report interval & Added some logs (#15905) (#15908)
* partial

* Changed default

* Update PipeDataNodeTaskAgent.java

* Next

+++ iotdb-core/node-commons/src/main/java/org/apache/iotdb/commons/pipe/config/PipeConfig.java
@@ -393,7 +393,7 @@ public class PipeConfig {
 
   /////////////////////////////// Logger ///////////////////////////////
 
-  public int getPipeMetaReportMaxLogNumPerRound() {
+  public double getPipeMetaReportMaxLogNumPerRound() {
     return COMMON_CONFIG.getPipeMetaReportMaxLogNumPerRound();
   }
